generando war con cambios de formato montos y comentado cce 3

+++ src/main/java/com/bancoexterior/app/inicio/service/GruposMenuServiceImpl.java
@@ -18,7 +18,8 @@ public class GruposMenuServiceImpl implements IGruposMenuService{
 	
 	@Override
 	public void borrarRealcion(GruposMenuPk id) {
-		repo.deleteById(id);
+		if(repo.existsById(id))
+			repo.deleteById(id);
 		
 	}
